+++ src/zope/pytestlayer/tests/fixture/single_layered_suite/test_core.py
@@ -1,0 +1,44 @@
+import doctest
+import unittest
+
+from plone.testing import layered
+
+from zope.pytestlayer.testing import log_to_terminal
+
+
+class FooLayer(object):
+
+    @classmethod
+    def setUp(cls):
+        cls.layer_foo = 'layer foo'
+
+    @classmethod
+    def tearDown(cls):
+        del cls.layer_foo
+
+    @classmethod
+    def testSetUp(cls):
+        log_to_terminal(cls.pytest_request, 'testSetUp foo')
+        cls.test_foo = 'test foo'
+
+    @classmethod
+    def testTearDown(cls):
+        log_to_terminal(cls.pytest_request, 'testTearDown foo')
+        del cls.test_foo
+
+
+def test_suite():
+    suite = unittest.TestSuite()
+
+    OPTIONFLAGS = (doctest.ELLIPSIS | doctest.NORMALIZE_WHITESPACE)
+
+    suite.addTests([
+        layered(
+            doctest.DocFileSuite(
+                'doctest.txt',
+                optionflags=OPTIONFLAGS,
+            ),
+            layer=FooLayer,
+        ),
+    ])
+    return suite